requirements.txt
streamlit
matplotlib
pandas
folium
streamlit_folium
branca
plotly.express
